Started adding functionality for a Fog of War effect.
Start the fundamentals of a Fog of War effect to go along with the Field of View effect from the previous commit. Seems to be compatibility issues with the Universal Render Pipeline used to create the darkness effect with the field of view. As a workaround is required to get both features working together this commit will be mainly a backup (for the working FOV and the half working FOW).