Adjusted Render Textures to display higher resolution and experimented with shaders.
The Render Textures now have a higher resolution, meaning the vison cone is a lot crisper and cleaner looking (Might still adjust further). Additionally, some Shader implementation was attempted for enemy rendering in the vision cone again, sadly is working in the scene but not in the game view yet, more research and experimentation needed.