Added a basic Inventory System based on the InventorySystem-FinalProject repository.
The inventory system from my InventorySystem-FinalProject repository has been added to the project and is fully functional. The point of the project was to be a universal inventory system, and while there were a few hiccups adding it to the current project, it has still lived up to its name. Other changes in this build were some more organazing of the project, as well as freeze functions to stop player, vision, guns and the arm from interaction during menus or other interactions.