Added Item pickup functionality, incompatible item flags for inventories and testing a new Unity Attribute for flags enumeration.
One of the main additions in this commit is the item pickup script. This script is based on the Interactable class and helps add items to the game world available for pickup by the player with interaction. Other changes in this commit are to the base inventory script. The base inventory now has an incompatible item check, used to block certain items from certain inventories based on selected enums in the Unity Inspector. These enums are using a new Unity Attribute called Flags Enumeration, which allows multiple choice when dealing with enums in the inspector (The Flags Enumeration is a test at the moment, so will be potentially removed if a better solution is discovered).