The weapon in an inventory is now selectable, as well as bullets de-spawning at a certain distance.
When a weapon is collected, it can now when equipped using the number keys corresponding to its slot. When a slot is selected that contains a weapon, the game will spawn a weapon in hand, making the weapon available for fire (There is no ammo requirement just yet). Along with this change the bullets have also had an update, no longer will they continue on forever if they miss a collision. A public distance value is now available on the bullet script to set a de-spawning range.