Gun is now set to selected slot straight away, the interactable indicator is only visible in the light, and various other changes.
A function to confirm the players InHand Object is now run every time a weapon or item pickup occurs. This way if a weapon pickup occurs and a slot has been selected the weapon will set to that slot and equipped weapon. Other changes to item interactions are the Interactable Visual Indicators, which now hide in the shroud and fog (As well as using the new Unity Text Mesh Pro component).

Ammunition functionality was also added in this commit, requiring bullet items in the player inventory to shoot a gun. This ammo must match the weapon attempting fire, as well as only using one item stack at a time. As well as ammunition, some additional sprites have been added to the project. These sprites will help clear up the inventory and world with two new sprites for use with the gun and ammo items.

Lastly, this commit adds functionality to toggle the Field Of View of the player, as well as moving all Field Of View adjustments to the FOV script. Hopefully, the ability to toggle FOV will become a fun feature for the future of the game or at the very least an element for different modes that don't require it.