Added support for the FOV system over the network.
FOV system is now fully functional over the network, however still having some trouble with the Fog Of War. But as for the FOV system, both players now have a light source that can see the other player and any enemies or items. On top of this, there is a new feature; each player can turn off their light and be unseen outside of the other players light source. Along with these changes, there are other small changes and bug fixes.