Committing some unfinished additions as I wanted to back up the project before updating unity.
Players now are assigned a random colour on connection to the game server, working currently, but permissions and security aren't how I would like it yet, so we will see if this feature stays around. Along with this change, the connection/pause menu can now be open again with a button press rather than using the unity editor.