Updated to the latest version of Unity and MLAPI, and now RPC's and NetworkVariables work correctly!
Updated Unity from 2020.1.6f1 to 2021.1.24f1. This update allowed for the install of the latest version of MLAPI. The reason this update is necessary is that RPC calls and NetworkVariables weren't working correctly, turns out because I was accidentally using the version before Unity took over development. So after everything was updated and code stitched back together, the script handling player colour changes is now working correctly and as intended!