Various Bug Fixes!
Fixed Error when Player disconnected that presented "Trying to destroy object 0" error code, fixed player collider, so it doesn't seem so "Clippy", and the Inner FOV/FOW object no longer goes through walls and is now using the FOV script. The final fix/change for this commit is now the FOV objects (Which renders the vision cone and inner-renderer) are a part of the Player Hierarchy. This new structure is important because it means the Player can spawn everything it needs and keep it contained to itself, much easier to move around and dispose of when disconnecting.