Started implementation of Item equipping and Shooting.
While it is currently working visually in the project, many issues are still present under the hood. I'm pushing these changes anyway because I am going away and will be working on the project remotely. Details on changes will be in the following commits and blog posts hopefully this week.