Equipping, shooting, dropping items, and plenty of bug fixes!
Plenty of bug fixes this build, the big one being sync issues when players join the game late. Equipping items is now working a lot better than before, and it seems to be fairly consistent, there are some warnings and errors on the server side, but those are handled by MLAPI almost immediately. Item Dropping was also added, activated by a middle mouse click on the item you wish to drop from the inventory, and it will drop at the player's feet. However, the main focus of this build was shooting over the network, which I attempted multiple techniques to solve, but finally, something stuck, and it's working quite well now. More details on this build are noted on the blog!