Just a small update today, mainly build of bug fixes and code maintenance, but also a fundamental change to item equipping.
Changed item equipping from heavy server reliance to being handled locally using the "Sync The Action, Not The Result" solution used for the shooting. So basically, a message is sent to the server to inform other players that an equipping action occurred, and those clients then update that result locally. The only other changes to this version were rewrites and the organisation of small sections of code.